requirements.txt
ccxt==4.2.65
diskcache==5.6.3
nsepython==2.8
plotly==5.22.0
psycopg2==2.9.9
pytest==8.3.2
python-dotenv==1.0.1
Requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
tqdm==4.66.1
yfinance==0.2.40
